ADD friends list
It may be a bit annoying not to be able to access a friend list. For
example, trying to retrieve a previous post or see if there are any new
comments on them is almost impossible (you have to wait for the same
person to post something new). Also, the home screen being blank most of
the time can be surprising for user.

To fix this, a Friend list has been implemented.

* it can be accessed from the menu bar
* it lists people you follow
* if there is nothing to show on home screen, a link will offer to go to
  friend list to see older posts

When a friend has unread posts, you will see a green border around their
avatar.